requirements.txt
streamlit
snowflake-connector-python
snowflake-snowpark-python
pandas
matplotlib
google-generativeai
python-dotenv